+++ info/asp-info.h
@@ -6,7 +6,6 @@
 #define ASP_INFO_080177a8_14ce_11ed_b65f_7328ac4c64a3_H
 
 #include <stdlib.h>
-#include <stdint.h>
 
 #ifdef __cplusplus
 extern "C" {
@@ -25,7 +24,7 @@ const char *AspRunResultToString(int result);
 AspSourceInfo *AspLoadSourceInfoFromFile(const char *fileName);
 AspSourceInfo *AspLoadSourceInfo(const char *data, size_t size);
 void AspUnloadSourceInfo(AspSourceInfo *);
-AspSourceLocation AspGetSourceLocation(const AspSourceInfo *, uint32_t pc);
+AspSourceLocation AspGetSourceLocation(const AspSourceInfo *, size_t pc);
 const char *AspGetSourceFileName(const AspSourceInfo *, unsigned index);
 
 #ifdef __cplusplus